I hate refactoring
