+++ code/camera/pom.xml
@@ -82,7 +82,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>org.example.cameraapi/com.nbicocchi.javafx.App</mainClass>
+                            <mainClass>org.example.cameraapi/com.nbicocchi.javafx.camera.App</mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
                             <jlinkImageName>app</jlinkImageName>
