+++ code/camera/src/main/java/com/nbicocchi/javafx/camera/common/AlertWindows.java
@@ -1,4 +1,4 @@
-package com.nbicocchi.javafx.common;
+package com.nbicocchi.javafx.camera.common;
 
 import javafx.scene.control.Alert;
