+++ code/camera/src/main/java/com/nbicocchi/javafx/camera/effects/LiveEffectsInterface.java
@@ -1,4 +1,4 @@
-package com.nbicocchi.javafx.effects;
+package com.nbicocchi.javafx.camera.effects;
 
 import javafx.scene.image.ImageView;
