+++ code/camera/src/main/java/module-info.java
@@ -7,12 +7,12 @@ module org.example.camera {
     requires javafx.base;
 
 
-    opens com.nbicocchi.javafx to javafx.fxml;
-    exports com.nbicocchi.javafx;
-    exports com.nbicocchi.javafx.common;
-    opens com.nbicocchi.javafx.common to javafx.fxml;
-    exports com.nbicocchi.javafx.effects;
-    opens com.nbicocchi.javafx.effects to javafx.fxml;
-    exports com.nbicocchi.javafx.controller;
-    opens com.nbicocchi.javafx.controller to javafx.fxml;
+    opens com.nbicocchi.javafx.camera to javafx.fxml;
+    exports com.nbicocchi.javafx.camera;
+    exports com.nbicocchi.javafx.camera.common;
+    opens com.nbicocchi.javafx.camera.common to javafx.fxml;
+    exports com.nbicocchi.javafx.camera.effects;
+    opens com.nbicocchi.javafx.camera.effects to javafx.fxml;
+    exports com.nbicocchi.javafx.camera.controller;
+    opens com.nbicocchi.javafx.camera.controller to javafx.fxml;
 }
