+++ code/camera/src/main/resources/com/nbicocchi/javafx/camera/controller/save-dialog.fxml
@@ -14,7 +14,7 @@
 <?import javafx.scene.layout.VBox?>
 
 <DialogPane maxHeight="-Infinity" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="com.nbicocchi.javafx.controller.DialogController">
+            fx:controller="com.nbicocchi.javafx.camera.controller.DialogController">
    <expandableContent>
       <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="200.0" prefWidth="320.0"/>
    </expandableContent>
