+++ code/camera/src/main/resources/com/nbicocchi/javafx/camera/errorImage/ErrorImage.png
