+++ code/camera/src/main/resources/com/nbicocchi/javafx/camera/home.fxml
@@ -15,7 +15,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<StackPane fx:id="stackPane" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.nbicocchi.javafx.controller.HomeController">
+<StackPane fx:id="stackPane" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.nbicocchi.javafx.camera.controller.HomeController">
    <children>
       <BorderPane>
          <bottom>
