+++ code/camera/src/main/resources/com/nbicocchi/javafx/camera/icons/CameraIcon.png
