--- code/camera/src/main/resources/com/nbicocchi/javafx/save-dialog.fxml
@@ -1,116 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.Cursor?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ButtonType?>
-<?import javafx.scene.control.ChoiceBox?>
-<?import javafx.scene.control.DialogPane?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.StackPane?>
-<?import javafx.scene.layout.VBox?>
-
-<DialogPane maxHeight="-Infinity" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.nbicocchi.javafx.controller.DialogController">
-    <expandableContent>
-      <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="200.0" prefWidth="320.0" />
-    </expandableContent>
-   <buttonTypes>
-      <ButtonType fx:constant="CANCEL" />
-      <ButtonType fx:constant="OK" />
-   </buttonTypes>
-   <graphic>
-      <StackPane fx:id="root">
-         <children>
-            <VBox>
-               <children>
-                  <StackPane VBox.vgrow="ALWAYS">
-                     <children>
-                        <Label text="Select Format" StackPane.alignment="CENTER_LEFT">
-                           <StackPane.margin>
-                              <Insets bottom="2.0" />
-                           </StackPane.margin>
-                        </Label>
-                     </children>
-                  </StackPane>
-                  <StackPane VBox.vgrow="ALWAYS">
-                     <children>
-                        <ChoiceBox fx:id="typeChoiceBox" StackPane.alignment="CENTER_LEFT">
-                           <StackPane.margin>
-                              <Insets bottom="10.0" />
-                           </StackPane.margin>
-                           <cursor>
-                              <Cursor fx:constant="HAND" />
-                           </cursor>
-                        </ChoiceBox>
-                     </children>
-                  </StackPane>
-                  <StackPane VBox.vgrow="ALWAYS">
-                     <children>
-                        <Label text="File name" StackPane.alignment="CENTER_LEFT">
-                           <StackPane.margin>
-                              <Insets bottom="2.0" />
-                           </StackPane.margin>
-                        </Label>
-                     </children>
-                  </StackPane>
-                  <StackPane VBox.vgrow="ALWAYS">
-                     <children>
-                        <TextField fx:id="fileNameTxtField" text="capture">
-                           <StackPane.margin>
-                              <Insets bottom="20.0" />
-                           </StackPane.margin>
-                           <cursor>
-                              <Cursor fx:constant="TEXT" />
-                           </cursor>
-                        </TextField>
-                     </children>
-                  </StackPane>
-                  <StackPane VBox.vgrow="ALWAYS">
-                     <children>
-                        <Button fx:id="selectDirButton" mnemonicParsing="false" onAction="#onSelectDir" text="Browse..." StackPane.alignment="CENTER_LEFT">
-                           <cursor>
-                              <Cursor fx:constant="HAND" />
-                           </cursor>
-                        </Button>
-                     </children>
-                     <VBox.margin>
-                        <Insets bottom="5.0" />
-                     </VBox.margin>
-                  </StackPane>
-                  <StackPane VBox.vgrow="ALWAYS">
-                     <children>
-                        <Label text="Selected Directory:" StackPane.alignment="CENTER_LEFT">
-                           <StackPane.margin>
-                              <Insets bottom="2.0" />
-                           </StackPane.margin>
-                        </Label>
-                     </children>
-                  </StackPane>
-                  <StackPane VBox.vgrow="ALWAYS">
-                     <children>
-                        <TextField fx:id="selectDirTxtField" mouseTransparent="true" />
-                     </children>
-                  </StackPane>
-               </children>
-               <padding>
-                  <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
-               </padding>
-            </VBox>
-         </children>
-      </StackPane>
-   </graphic>
-   <content>
-      <VBox>
-         <children>
-            <StackPane VBox.vgrow="ALWAYS">
-               <children>
-                  <ImageView fx:id="preview" fitHeight="150.0" fitWidth="200.0" pickOnBounds="true" preserveRatio="true" StackPane.alignment="CENTER" />
-               </children>
-            </StackPane>
-         </children>
-      </VBox>
-   </content>
-</DialogPane>
